--- Projektarbeit III/Assets/Scripts/UI & Inspector/FramesPerSecond.cs
@@ -1,26 +1,0 @@
-using UnityEngine;
-
-public class FramesPerSecond : MonoBehaviour
-{
-    private float deltaTime = 0.0f;
-
-    void Update()
-    {
-        deltaTime += (Time.unscaledDeltaTime - deltaTime) * 0.1f;
-    }
-
-    void OnGUI()
-    {
-        int width = Screen.width, height = Screen.height;
-        GUIStyle style = new GUIStyle();
-
-        Rect rect = new Rect(0, 0, width, height * 2 / 100);
-        style.alignment = TextAnchor.UpperLeft;
-        style.fontSize = height * 2 / 100;
-        style.normal.textColor = Color.white;
-        float msec = deltaTime * 1000.0f;
-        float fps = 1.0f / deltaTime;
-        string text = string.Format("{0:0.} fps", fps);
-        GUI.Label(rect, text, style);
-    }
-}
